+++ devel/sclm/sclm.cpp
@@ -1006,30 +1006,6 @@ sclr::registry_ &sclm::GetRWRegistry( void )
 	return sclr::GetRWCommonRegistry();
 }
 
-flf::rWFlow &sclm::rWFlowRack::Init( const fnm::name___ &FileName )
-{
-	_FileName.Init( FileName );
-
-    sclm::CreateBackupFile( _FileName );
-    _BackedUp = true;
-
-    if ( _Flow.Init( _FileName ) != tol::rSuccess )
-        sclm::ReportFileOpeningErrorAndAbort( _FileName );
-
-    return _Flow;
-}
-
-txf::text_oflow__ &sclm::text_oflow_rack___::Init( const fnm::name___ &FileName )
-{
-	if ( FileName.IsEmpty() ) {
-		return cio::COut;
-	} else {
-		_TFlow.Init( rWFlowRack::Init( FileName ) );
-
-		return _TFlow;
-	}
-}
-
 namespace {
 	// Common registry locked upstream.
 	void DumpRegistry_(
@@ -1119,12 +1095,6 @@ qRE
 }
 
 #undef T
-
-void sclm::rWFlowRack::HandleError( void )
-{
-	if ( _BackedUp )
-		sclm::RecoverBackupFile( _FileName );
-}
 
 namespace {
 	 inline void FillPluginRelatedTags_(