+++ tools/tasq/tasqxml.h
@@ -26,12 +26,32 @@
 # include "xml.h"
 
 namespace tasqxml {
+  qENUM( Filter ) {
+    fDescription_,
+    fId_,
+    f_amount_,
+    f_Undefined_,
+  };
+
+# define FF_( name )	ff##name = ( 1 << f##name##_ )
+
+  qENUM( FilterFlag ) {
+    FF_( Description ),
+    FF_( Id ),
+    ffAll = ( ( 1 << f_amount_ ) - 1 ),
+    ffExport = ffDescription,
+    ffDisplay = ffId
+  };
+
+# undef TF_
+
   void Write(
     const tasqtasks::dBundle &Bundle,
+    int TokenFlags,
     xml::rWriter &Writer);
 
   void Parse(
-    xtf::sRFlow &Flow,
+    xml::rParser &Parser,
     tasqtasks::dBundle &Bundle);
 }